Add type hints for a number of test helper functions
This adds type hints for all `pytest` fixtures in
`sybil/` and `tests/`, as well as some other test helper functions.

Some people prefer to use `Iterator` over `Generator` for typing
`pytest` fixtures, but this can cause issues, e.g.
https://youtrack.jetbrains.com/issue/PY-40318/pytest-fixtures-with-yield-return-types-defined-as-Iterator
and Generator is an "accepted answer" on `pytest-dev`:
https://github.com/pytest-dev/pytest/discussions/7809#discussioncomment-87250.
I don't fully understand that, but I went with the accepted answer.