Avoid using private variable _unittest_reportflags from doctest
Using private variables from a dependency increases risk of undesired behaviour change.
mypy does not like the use of private variables from dependencies.

This change removes the use of _unittest_reportflags.
Instead, it changes the code to get the old flags, by setting them to 0 (arbitrary),
and then it sets them back.

Reading the code makes me think that this is a valid change, however,
there is some risk as the tests are not thorough: even if I remove
the line:

```
optionflags |= old_doctest_unittest_reportflags
```

the tests still pass.