Add a number of "assert <> is not None" for mypy
This codebase uses multiple methods and properties which are typed as Optional,
and makes the assumption that they are not None.

There are multiple ways of getting around this for the type checker.
For example:

* `type: ignore` comments. A previous review comment explained that this is not wanted.
* `cast`.  A previous review comment explained that this is not wanted.
* `assert <> is not None`.  A previous review comment explained that this is not wanted.

I chose the latter as it is most to my taste.
I am open to specific suggestions for alternatives.