Use specific codes for type: ignore comments
I recommend this because it reduces the chance of accidentally
ignoring a valid type error that you did not mean to ignore.

I'll also point out while here that the type errors here are because the stubs
for `doctest` are incomplete, and that a "real" fix would be to upstream some
changes to `typeshed`.
That was on my list but if you're happy with `ignore`, that's also good.